:sparkles: Various upgrades, converted to library
