+++ README.md
@@ -1,2 +1,8 @@
 # dolreader
 A python library for reading data from and writing data to Dol files (GC/Wii executables).
+
+# Features
+- Complete file support
+- Sections and saved DOLs are properly aligned
+- Easy branch insertion
+- Sections are abstracted from the file, making access easy and efficient
