+++ dolreader/__init__.py
@@ -1,0 +1,3 @@
+__version__ = "1.0.2"
+__author__ = 'JoshuaMK'
+__credits__ = 'RenolY2: https://github.com/RenolY2'
