+++ dolreader/exceptions.py
@@ -1,0 +1,4 @@
+class UnmappedAddressError(Exception): pass
+class SectionCountFullError(Exception): pass
+class AddressOutOfRangeError(Exception): pass
+class IncompleteSectionError(Exception): pass
